perf: 修改收集用户中间
