+++ apps/accounts/models/__init__.py
@@ -1,4 +1,3 @@
 from .base import *
 from .account import *
 from .automations import *
-from .gathered_account import *
